requirements.txt
Django==5.0.2
django-rest-framework==0.1.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
beautifulsoup4==4.12.3
pymongo==4.6.1
mysqlclient